Fix several Java compiler warnings in autofocus plugins
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@12282 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ autofocus/AutofocusDuo.java
@@ -153,7 +153,7 @@ public class AutofocusDuo extends AutofocusBase implements Autofocus  {
       // use default dialog
             
       String afDevices[] = MMStudioPlugin.getAutofocusManager().getAfDevices();
-      String allowedAfDevices[] = new String[(int)afDevices.length - 1];
+      String allowedAfDevices[] = new String[afDevices.length - 1];
 
       try {
          PropertyItem p1 = getProperty(KEY_AUTOFOCUS1);
